Break up NodeWidget into smaller components
