+++ src/core/components/widgets/Node/index.ts
@@ -1,0 +1,3 @@
+import NodeWidget from './Node';
+
+export default NodeWidget;
